+++ Spring/src/main/java/bitc/full502/final_project_team1/core/domain/repository/projection/BuildingListProjection.java
@@ -6,7 +6,7 @@ public interface BuildingListProjection {
     String getRoadAddress();
 
     Integer getAssigned();          // 1 or 0 (CASE WHEN …)
-    Integer getAssignedUserId();
+    Long getAssignedUserId();
     String  getAssignedUserName();
 
     Long    getResultId();
